pyproject.toml
[tool.poetry]
name = "dags"
version = "0.1.0"
description = ""
authors = ["ranog <jprnogueira@yahoo.com.br>"]
readme = "README.md"
packages = [{ include = "dags", from = "." }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
poetry-plugin-export = "^1.9.0"
apache-airflow = "^3.0.1"
pandas = "^2.2.3"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pre-commit = "^4.2.0"
ruff = "^0.11.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]
line-length = 100
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]
select = [
    "F",  # Pyflakes
    "E",  # Pylint
    "W",  # PEP 8
    "C",  # Complexity
    "T",  # Type hints
    "I",  # Imports
]  

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
